changed the type of image

+++ line.py
@@ -145,7 +145,7 @@ def get_edges_on_lines(edges, img_w_lines, kernel_size=3):
     """
     final = np.zeros(edges.shape)
     r = widen_lines(img_w_lines, kernel_size=kernel_size)
-    
+
 
     for i in range(edges.shape[0]):
         for j in range(edges.shape[1]):
@@ -171,10 +171,9 @@ def widen_lines(img, kernel_size=3):
     ------
     The image with widened lines
     """
-
     kernel = np.ones((kernel_size, kernel_size))
     r = cv2.filter2D(img, -1, kernel, borderType=cv2.BORDER_CONSTANT)
-    r = np.where(r > 0, 255.0, 0)
+    r = np.where(r > 0, 255, 0).astype(np.uint8)
 
     return r
